+++ packages/design-system-angular/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.15.1-pr-90.1 (2023-09-14)
+
+**Note:** Version bump only for package @geovistory/design-system-angular
+
+
+
+
+
+## 1.15.1-pr-90.0 (2023-09-08)
+
+**Note:** Version bump only for package @geovistory/design-system-angular
+
+
+
+
+
 ## 1.15.1-pr-85.0 (2023-09-07)
 
 **Note:** Version bump only for package @geovistory/design-system-angular